backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
redis>=5.0.1
celery>=5.3.6
docker>=7.0.0
httpx>=0.26.0
pydantic-settings>=2.0.0
openai>=1.68.2
pgvector>=0.3.6
pypdf>=5.4.0
python-pptx>=1.0.2
tiktoken>=0.9.0
